refactor: 탑승정보 선택 패널 파일 구조 수정
