+++ frontend/react-app/src/pages/home/components/panelItem/schedule/SchedulePanelItem.tsx
